milestone 6 mvp

+++ apps/api/src/authorizations/deleteAuthorizations.ts
@@ -1,8 +1,17 @@
 import { pool } from "../db/pool";
+import { isOwner } from "../middleware/roleValidation";
 
 export async function deleteAuthorizations(_req: any, _res: any) {
 	const userId = _req.query.userId;
 	const projectId = _req.query.projectId;
+
+	if (_req.user.role !== "admin" && !(await isOwner(_req.user.sub, projectId))) {
+		return _res.status(403).json({
+			error: "Forbidden",
+			message: `This action requires one of these roles: admin, owner (of project).`
+		});
+	}
+
 	try {
 		const response = await pool.query(`DELETE FROM authorizations WHERE user_id=$1 AND project_id=$2`, [userId, projectId]);
