+++ apps/api/src/users/deleteUsers.ts
@@ -2,6 +2,14 @@ import { pool } from "../db/pool";
 
 export async function deleteUsers(_req: any, _res: any) {
 	const id = _req.params.id;
+
+	if (_req.user.role !== "admin" && _req.user.sub !== id) {
+		return _res.status(403).json({
+			error: "Forbidden",
+			message: `This action requires one of these roles: admin, self==id.`
+		});
+	}
+
 	try {
 		const response = await pool.query(`DELETE FROM users WHERE id=$1`, [id]);
 		await pool.query(`DELETE FROM authorizations WHERE user_id=$1`, [id]);
